refactor(portfolio/data/dataApi.ts):
rename projects.ts to dataApi.ts for clarity
